coral: add 2S batteries and support default battery from 2S or 3S one.
In order to support 2S and 3S batteries in the same time, we need a way
to figure out whether default battery type should be came from the group of
2S or 3S batteries. This is in order to set PPVAR_SYS correctly from
charger driver when battery is not connected.

BUG=b:70701623
Signed-off-by: Marco Chen <marcochen@chromium.org>
BRANCH=coral
TEST=manually test in the device with 3S battery. For 2S battery, needs
to wait for the proto board of babymako.

Change-Id: I9c576507b7d7cd83110f68c38c3c342144d78ed8
Reviewed-on: https://chromium-review.googlesource.com/948222
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Commit-Queue: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>