Nami: Disable one USB-C port
The TCPC of the left USB-C port is being swapped with Anx7447. This
patch temporarily disables the port so that the boot is not inhibited
by running a wrong driver on the port.

BUG=b:73793947
BRANCH=none
TEST=Boot Nami with/without battery. Verify it can be charged through
the right port.

Change-Id: I30b63f87cb3669c38a04e908b6556eef1424fbd7
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/957886
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Trybot-Ready: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>