meowth: zoombini: Ignore AC when checking cutoff.
When charge manager safe mode is enabled, it prevents us from selecting
CHARGE_PORT_NONE if we have a battery which is disconnected/cutoff.  The
cutoff reporting code was essentially making the assumption that if AC
is not present, then we must be running off of battery power.  However,
AC presence is debounced and can take some time to be reported to the
system.  Therefore, we may wrongly assume that we are running off of
battery power and cut off our AC power source.

This commit simply removes that assumption and still consults the
battery regarding the cutoff/disconnect state.

BUG=b:74125001
BRANCH=stabilize-meowth-10444.B
TEST=flash meowth; cutoff battery, insert AC, verify system comes up.
Verify that CHARGE_PORT_NONE is not selected until battery wakes up.

Change-Id: I79a27bd2d5f09c2ffe65df98402c0ae0182fdba2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/958101
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Shen-En Shih <petershih@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>