USB-PD: Avoid discovering identity twice on boot
Some monitors do not like the same mode to be entered twice. They
refuse to show any picture when it happens.

This patch makes the TCPM check whether the port is a UFP or not before
setting PD_FLAGS_CHECK_IDENTITY flag when transitioning from S5 to S3.
If not (port is a DFP), it skips setting the flag.

Key observations are:
- Setting CHECK_IDENTITY there was useful for the following reasons:
  * In S5 we are in FORCE_SINK. So we might not be able to discover mode
    (due to the fact that we are a UFP initially when being a SINK).
  * Once we start, we want to discover the full capabilities.
- The use case above is not true if we are a DFP (since we can
  discover mode immediately)
- So it's useful IFF we are a UFP at transition.

BUG=b:75958206
BRANCH=none
TEST=Verified with a reworked Fizz on USB-C monitor/adapter
(LG 27UD88, HP S240n, Dingdong, Hoho) as follows:
1. Plug-in boot
2. Manual recovery
3. Servo reset
4. EC reboot command
5. Warm reboot
6. Power button boot from S5
7. Suspend / resume

Change-Id: I97a3fa671582552c029f7da140b55e7724b56ecc
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/963844