Snappy: Don't discharge on AC if system is off
When an AC charger is plugged in, the EC enable discharge-on-ac:
stopping current from VBUS and draw current from the battery. This
feature is introduced to avoid drawing too much current at once from
an AC adapter. (CL:431138)

We suspect this feature causes the system to choke when the battery is
deeply discharged.

This patch disables discharge-on-ac if the system is in G3 or S5.

BUG=b:35584495,b:74586187
BRANCH=reef
TEST=none

Change-Id: I2ddee48d4bbaa448b3d1f77b5ab079d76a4be8cb
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1011594
Reviewed-by: Saurabh Madan <ssmadan@chromium.org>