Fizz: Configure USB_C0_PD_RST_ODL as GPIO_ODR_HIGH
USB_C0_PD_RST_ODL is connected to PS8751 reset pin and currently configured
as GPIO_OUT_LOW. This causes the EC to generate an unintentional reset
pulse when Cr50 resets the EC on power-on reset. When PS8751 is in reset,
the CC line is pulled low. This unintentional reset lasts longer than
tPDDebounce ("CC Timing" in the USB Type-C spec), thus causes Z32 to bring
down VBUS and the system to brown out.

This patch configures the pin as GPIO_ODR_HIGH. This will remove the
unintended reset pulse. This pin will continue to be used to generate a
reset pulse as it has been doing without exceeding tPDDebounce.

BUG=b:74354232
BRANCH=none
TEST=Verify Fizz boots on HP Z32 by:
1. Cold reboot (by servo) and
2. Plug-in USB-C cable (connected to Z32)

Change-Id: Ieae7c25e8261cf3b708e76c8340b07af5a61720f
Reviewed-on: https://chromium-review.googlesource.com/1031736
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>