nautilus: Only disable keyscanning when not in S0.
Devices should not disable keyboard scanning when the chipset is in S0.
Thy should keep the keyboard enabled and continue to forward the keys to
the AP such that the AP can make the decision whether to ignore them or
not.

BUG=b:79670788
BRANCH=poppy
TEST=Flash nautilus; Connect external display, reduce brightness on
internal display to make external display the primary display; Enter
tablet mode, verify that keyboard presses still show up in the UI.

Change-Id: I452299726c6e8b388eb1061c23bed013228c2a2c
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1089823
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>