cheza: Fix USB peripheral mode when using a Type-C to std-A male cable
In the PD state machine, the decision is that if the data is DFP, then
connect the mux, but if the data role is UFP, then disconnect the mux.
To make the Type-C to std-A male cable work, the UFP should also
connect, need to disable the CONFIG_USBC_SS_MUX_DFP_ONLY.

BRANCH=none
BUG=b:110265762
TEST=Used an image enabling USB peripheral mode and USB gadget zero,
typed EC command to mux the port-1 HS to AP, plugged a Type-C to std-A
male cable, the host side can enumerate the gadget zero device and
the dmesg log shows the detection too.

Change-Id: I532e53bd0832c50a86a91d1c8d84fd55af543eca
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1102985
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>