DragonEgg: Initial skeleton for DragonEgg
This CL adds DragonEgg in /board and /baseboard. Only minimal gpio
signals are defined to allow successful build. The /baseboard files
are currently empty, but the plan is to use baseboard for functions
that will be common for ite EC and ILK.

BRANCH=none
BUG=b:110880394
TEST=make buildall

Change-Id: Ia018692f277efaceef85a060b6585accc82fddfd
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1117133
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>