core/host: Fall back to udelay when task is invalid
When running fuzzing tests, the sanitizer library may call usleep
from the main thread, and our implementation thinks that usleep
is called from idle task (task_id == 0), and just waits for an
event that will never arrive.

Make sure the default task id is invalid, and fall back to udelay
if we are in an invalid task.

BRANCH=none
BUG=chromium:854975
TEST=Fuzzing tests do not fail with strange errors.

Change-Id: Icc3fdce30b54dfb06913a3d6cbabaa07e1266ba6
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1116623
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>