USB PD: Don't restore the default power state.
Now, the PD state machine will try and issue a Soft Reset if the port
was previously in an explicit contract prior to the EC resetting.
However, there was a bug where the state machine would actually restore
the default power role instead of the actual power role that was stored
in in battery-backed RAM.

This commit fixes that bug by simply setting the default power role if
and only if it hasn't been set already by going down the Soft Reset code
path.

BUG=None
BRANCH=None
TEST=flash nocturne, verify that soft reset path works.

Change-Id: Ia2baeda2674c6febf4c8c04496a131698099dcdc
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1117368
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>