tcpc: disallow PD_SUSPEND if not enough power
We do not want to allow depthcharge to suspend the PD port for a
firmware upgrade if there is not enough battery power to last us during
the full firmware upgrade.

We cannot use the active_charger function as the EC has just jumped to
RW and has not fully initialized the charger state. We use Vbus presence
as a good enough approximation that power may be coming in from that
port.

BRANCH=none
BUG=b:78334391
TEST=No battery or deeply discharged battery will skip TCPC upgrade on a
yorp if there is a usb charger in that port. Otherwise the upgrade will
still be attempted.

Change-Id: I9af9cb84cfae611b6ab051d318e2083962424de1
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1123314
Reviewed-by: Furquan Shaikh <furquan@chromium.org>