atlas: tune MP2949 PMIC on boot
we need to tune some of the MP2949 PMIC parameters for stability.  we
can do this from the EC when we boot the AP.  at this point we always
check to see if the desired values are programmed and apply them if
necessary.  eventually, the PMIC will be pre-programmed and we won't
have to do this any more.

BUG=b:110890675
BRANCH=none
TEST=verified settings are being applied

Change-Id: Ia814a362934f9cb9eb5c28dc0588f7425b3ab79e
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1124219
Commit-Queue: Caveh Jalali <caveh@google.com>
Tested-by: Caveh Jalali <caveh@google.com>
Reviewed-by: Caveh Jalali <caveh@google.com>
Trybot-Ready: Caveh Jalali <caveh@google.com>