nocturne: Adjust IMVP8 step decay quantization.
The IMVP8 vendor suggested to change the step decay quantization from 3
to 1.  Doing this seems to help with the hangs that were seen.

This commit simply has the EC write to the IMVP8 to set this register
~250ms after sequencing to S0.

BUG=b:111224125
BRANCH=master
TEST=Flash nocturne, boot to S0, read back register 0xFA from IMVP8,
verify that it reads back as 0x0ac5.

Change-Id: Ic7444c763ed4da3a4b80b3a9e79b60c5fa984345
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1132718
Reviewed-by: Nick Vaccaro <nvaccaro@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>