flash-stm32h7: Invalidate D-cache after erase/write operations
When erasing or writing to the flash, the core may still have
some of the old data in cache. Among other things, this confuses
flashrom when it tries to read back an erased block, as it thinks
that the erase operation was unsuccessful.

This, of course, will have some moderate performance impact when
doing memory operations, but hopefully those are rare enough to
not matter significantly.

The I-cache does not have such issues, as we invalidate it on
sysjump/boot, when it actually matters.

BRANCH=none
BUG=b:111144409
TEST=flashrom works reliably:
 cp ec.bin ec.bin.blank
 dd if=/dev/zero of=ec.bin.blank bs=1024 seek=1024 count=1024 conv=notrunc
 flashrom -V -p ec:type=fp -i EC_RW -w ec.bin.blank
 flashrom -V -p ec:type=fp -i EC_RW -w ec.bin
TEST=flashread returns correct data after an erase
EC> flashwrite 0x180000 1024 => Initialize flash
EC> flashread 0x180000 1024 => Loads data in cache
EC> flasherase 0x180000 131072
EC> flashread 0x180000 1024 => Make sure cache was invalidated

Change-Id: I58c26aa5a5715cea4fd7997414fd3f0c8a784572
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1134719
Reviewed-by: Randall Spangler <rspangler@chromium.org>