USB PD: Defer entering LPM by 5 seconds at init.
When the EC boots up from a charger being plugged in, the TCPC may
briefly report that there is nothing connected to it via sampling its CC
lines.  Our PD state machine would decide to put the TCPC into "low
power mode" by setting the dual-role auto toggle functionality.  The
port is deemed not connected at this point and shortly after, charge
manager may choose a charge port of "none" and cut off the VBUS coming
into the system.  This is especially problematic in the deeply
discharged battery case as it would result in a charger being physically
plugged into the port, but not actually charging the battery.

This commit simply adds a delay of 5 seconds before allowing a port to
enter it's auto toggle mode.  This gives the state machine a chance to
continuously poll the CC lines during which the TCPC actually reports
that there is a partner plugged into the port and begin the discovery.

BUG=b:112954174
BRANCH=nocturne
TEST=Flash nocturne; cutoff battery, plug in AC, verify that VBUS is not
dropped from the system.

Change-Id: I7cb897ec442fd42b96a397f74bbc206a705d2df8
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1187342
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>