FIXUP: driver: bmi160: harden interrupt and fifo processing
Actually 5ms is needed to reset interrupt of activity like significant
motion.
When processing FIFO, the interrupt will go down once the FIFO is
processed, but for activity, there is no easy way for the bit to go
down.
See
https://electronics.stackexchange.com/questions/299188/how-to-program-bmi160-inertial-measuremen-unit-to-latch-detection-flags

BUG=b:73557414,b:80284952
BRANCH=scarlet,poppy,nocturne
TEST=Without this change, the interrupt bitfield in the the interrupt
routine never goes back to 0.
Check the sensor is not getting stuck.

Change-Id: I986547787183377938ab08afe52fb2871d5c7c71
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1205203
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>