usb_i2c: provide a way to communicate non-i2c commands
On some occasions there might be a need to request an action from the
DUT which is related to i2c operations but can not be performed
through an i2c transaction. For instance when one need to generate the
sync sequence on the EC i2c interface before the EC can be programmed.

This patch introduces a facility which allows to register a handler
for commands sent on this special address which is picked at 0x78,
which becomes 0xf0 during write transaction address cycle.

BRANCH=cr50, cr50-mp
BUG=b:75976718
TEST=tested along with the rest of the patches to trigger generating
     of the ITE EC debugger sync sequence.

Change-Id: I269d4b8073b0d02f96ca526f221c02c38a036f3d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1214542
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1309041
Reviewed-by: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>