Make USB_PD_POWER_INFO detect inactive supplier
Currently, CMD_USB_PD_POWER_INFO does not report a supplier if the
available current is 0. This causes a dedicated charger (a barrel
jack adapter) to be ignored by the command if it's not the active
supplier.

This change allows a supplier which has 0 current to be reported
by the command as long as there is no other supplier with positive
current on the same port. With this change, the host can detect
a BJ adapter's physical presence regardless of its charging status.

In other words, a supplier with positive current always supersedes
whatever supplier with zero current.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chromium:841944,b:111667665,b:111777351
BRANCH=none
TEST=Verify ectool usbpdpower 1 return 'Port 1: Disconnected' and
'Port 1: SNK (not charging)' without and with a BJ adapter connected
respectively on Fizz.

Change-Id: I22ba5f147209177c0c9cac87a123d4c3d0f2ec68
Reviewed-on: https://chromium-review.googlesource.com/1236850
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>