cr50: use the latest Cr50 signer
Update make files to work with the latest version of the Cr50 signer.

Technically this is not necessary in the chip factory branch, as
signing does not have to happen in the branch, but it helps when a
board locked image from this branch needs to be created.

BRANCH=none
BUG=none
TEST=successfully created node locked images from this branch using
     H1_DEVIDS=... make variable.

Change-Id: I0fe37bd66b1f2fb995ebd022ca0fd4e1eb46299f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1236919
Reviewed-by: Randall Spangler <rspangler@chromium.org>