fpsensor: FP_INFO reports the template version
Add a field to the ec_response_fp_info structure to report the version
of the template format supported by the hardware.
We'd normally uprev the structure version, but given there are only 2
clients of that API (ectool and biod), we'll rather synchronise the
submission of the CLs to avoid unnecessary complexity.

BRANCH=nocturne
BUG=b:116979455
TEST=ectool --name=cros_fp fpinfo
TEST=start biod, log shows the format version

Change-Id: Ibbf2a3603d4f28c7550523f785aeb7fb740ac9ea
Signed-off-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1252460