chip/stm32/usb_hid_keyboard: Reload keyboard endpoint on reset
After the USB device is reset, force sending the keyboard data at
least once to make sure the host is aware of the curreny key/switches
status.

In particular, this is necessary for the host to be aware that we
are in tablet mode on host boot (this is not a problem on attach, as
other routines in the EC init sequence cause the endpoint to be
loaded anyway).

BRANCH=nocturne
BUG=b:117439202
TEST=Reboot nocturne with whiskers attached in tablet mode, see
     that nocturne is still in tablet mode after reboot:
     evtest --query /dev/input/event6 EV_SW SW_TABLET_MODE; echo $?
     => 10

Change-Id: I4b0e20a9f78c5262d7effd8bd37ace0d033b3f5a
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1270297
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 399bc14a001e2e12c158233081948f10ef885f6d)
Reviewed-on: https://chromium-review.googlesource.com/c/1276286