tcpm: cache the firmware version for anx74xx and ps8xxx chips.
All chip info for the anx74xx and ps8xxx chips should be cached
after the first read, so that the info can be accessed in the
future without worrying about chip state. Now that we check the
version registers for these two chips, we need to cache that
information as well.

BRANCH=none
BUG=b:117301622
TEST=on warm reboot with GBB flags cleared, cbmem -t shows
the EC verification stage taking ~30ms instead of ~65ms.

Change-Id: I6077438b58e64b66a23db2992ba025ae3a864c19
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1291794
Commit-Ready: Martin Roth <martinroth@chromium.org>
Reviewed-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1299896
Reviewed-by: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>