npcx: gpio: support internal pull-down when enable low voltage mode
There is the limitation that internal pull-up must be disabled when
a GPIO is configured in low voltage level. However, thers is no such
limitation of internal pull-down. The current GPIO driver disable no
matter pull-up or pull-down when low voltage mode is set.
This CL fixes it by:
1. enable internal PD when low voltage mode is set.
2. print warning message in the UART console when both low voltage and
internal PU flags are set for any GPIO defined in gpio.inc.

BRANCH=none
BUG=b:118339468
TEST=No build error for make buildall
TEST=define a gpio with internal PD+low-voltage in npcx7_evb/gpio.inc,
check the releated bits of PxPULL and PxPUD are set;
TEST=define a gpio with internal PU+low-voltage, check the warning
message is printed on the console.

Change-Id: I8e15125d3a2ccc73f84b8a559d12644b1d1af5f9
Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Reviewed-on: https://chromium-review.googlesource.com/1297872
Commit-Ready: CH Lin <chlin56@nuvoton.com>
Tested-by: CH Lin <chlin56@nuvoton.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1309044
Reviewed-by: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>