mkbp_event: deassert host irq when if the current event cannot be fetched
On boot, it is possible to encounter a case where a host event bit
is set, but the FIFO is actually empty (e.g. if an event was sent
when the AP is sleeping).

In this case, we should still deassert interrupt to AP, if there
are no other events pending.

BRANCH=none
BUG=b:117406392
TEST=Boot kukui, no "HC 0x67 err 9"

Change-Id: I5c23b3a37c16fce673478f863b6b7660aba59ee9
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1314308
Reviewed-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1484638
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>