Reland "Battery: USE Host full capacity to compute display percentage"
The full capacity known by the host may slightly differ from the full
capacity known by the EC because we notify the host of the new
capacity only if the difference is larger than 5 mAh.

This patch makes the EC use the host's full capacity instead of the
local full capacity to compute the display percentage.

(Relanding CL:1330101 patchset 6)

BRANCH=none
TEST=Verify display percentages printed by EC and power_supply_info
move up synchronously on charge and the LED and the taskbar icon
turn to full at the same time.
TEST=buildall

Change-Id: If9e0c9989e19bb9085866a86b47d1c200fc70b80
Reviewed-on: https://chromium-review.googlesource.com/c/1340542
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>