ec_commands: move __ec_* macros to end of struct definition
This matches the typical style for kernel structs, and it avoids
tripping up the kernel's kernel-doc parser/validator. This matters,
because we sync this header with the Linux kernel tree regularly.

This transformation was done entirely with vim macros, to match '^struct
__ec_' and move the second word down to the next '^};'.

BRANCH=none
BUG=chromium:621123
TEST=presubmits

Change-Id: I54c1b5b2daa47a54727049551ef1401a4881e1a5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1335685
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1603114
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>