PD 3.0: Update message type field size
For PD 3.0, bit 4 of the message header was added to the message type
field.  It should be safe to check that bit in all of our header type
checks regardless of version, since in PD 2.0 that field was
"Reserved/shall be set to 0".

BUG=None
BRANCH=None
TEST=builds, ampton PD negotiation works

Change-Id: I374c3d71210652a138e2ad53bd12e4b31eae9b0c
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1363814
Reviewed-by: Jett Rink <jettrink@chromium.org>
(cherry picked from commit 011a5575405bcc1f6f41e82eb5a896265b019467)
Reviewed-on: https://chromium-review.googlesource.com/c/1368504