nocturne: Raise ATTACH_MAX_MV to 900.
There are some devices where when connected to whiskers, the base attach
ADC reading is right on the threshold for max.  This may cause whiskers
to not be powered appearing as the keyboard not working.  This commit
simply increases the threshold to 900mV to allow more whiskers to be
detected.

BUG=b:122361740
BRANCH=firmware-nocturne-10984.B
TEST=`make -j buildall`

Change-Id: I10e2d58e5277abdd28a50a3aca3ef0f26d93aeba
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1399186
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>