atlas: only apply internal EC_WP_L pullup when needed
we had applied an internal pullup on the EC side to the EC_WP_L signal
to work around a problem.  we now have a board level fix on newer
boards, so we should only apply this workaround on older boards.

BUG=b:122568899,b:116670191
BRANCH=stabilize-atlas-11512.6.B
TEST=verified older boards still read EC_WP_L as high

Change-Id: Ia95b8649eed240a7c97c6b1cc20ed5ccea40e4f2
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1407244
Commit-Queue: Caveh Jalali <caveh@google.com>
Tested-by: Caveh Jalali <caveh@google.com>
Reviewed-by: Caveh Jalali <caveh@google.com>