coral:enable battery cutoff when shutdown_on_critical_battery
The machine can't power on when the soc of battery is 0% with AC adapter
plugging in and this behavior is not consistent as usual. The power
consumption is very low when battery is cutoff, so let battery cutoff
instead of EC hibernate to try to avoid battery soc dropping to 0%.
However, it would take more than 1 hour to consume 1% battery power
in G3 state. we also need prevent from hibernating in chipset_task
if battery soc <= BATTERY_LEVEL_SHUTDOWN to get chance to cutoff the
battery.

BRANCH=coral
BUG=b:120407297
TEST=Discharge the battery to verify if the battery is cutoff.

Change-Id: Ia8ce02e9e6e9b1a0eafe872664fd246667a18b00
Signed-off-by: lu.zhang <lu.zhang@bitland.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1408749
Reviewed-by: Scott Collyer <scollyer@chromium.org>