LSM6: Correct timestamp reporting and FIFO length
When an accelerometer sends a sample to the motion_sense task with a
timestamp, the timestamp should represent the time of the IRQ which
caused the FIFO read.  This change caches a timestamp after the
accelerometer data is read to use when pushing all data from that read.

This also corrects the number of bits in the FIFO length field.

BRANCH=octopus
BUG=b:120679547
TEST=builds, phaser360 CTS pass rate unaffected

Change-Id: I220aa2e8fa23af3f7833999cdfac966e8695c831
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1413670
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1719398
Reviewed-by: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>