motion_sense: send ODR events to the AP
This commits enables the EC to send a motion sense event to
the AP when a sensor's ODR is changed.

This is done in a way that is backwards compatible on the AP side,
in order to enable previous kernels to continue to work
(albeit at a loss of new functionality) with up-to-date ECs.

Since clients are not required to use a FLUSH when changing ODR, it
is possible for the sensor ring to get in a state where it is spreading
based on out-of-date delay information, which we observed leading to
spreading of overly large intervals, and thus too far into the future.

This patch ameliorates that by causing an ODR change to implicitly flush
the batching history.

BUG=b:123700100
TEST=run CTS sensor tests, observe the sensor ring not
     spreading too far into the future due to a period
     change
BRANCH=bobba

Change-Id: I5a881e9fc5ddb1dbe6690c455d43dc20bffe889a
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1450288
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1719537
Reviewed-by: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>