pd: Don't Try.Src if battery is disconnected.
If a battery is disconnected (the discharge FET is disabled) we should
not attempt Try.Src even though the battery state of charge may be above
CONFIG_USB_PD_TRY_SRC_MIN_BATT_SOC.  The reason is that in this case,
the battery is not providing charge to the system, so attempting Try.Src
will temporarily cut VBUS which may be the only power source in the
system at this time.

This commit simply checks the battery disconnect state when deciding to
attempt Try.Src.

BUG=b:123722302
BRANCH=firmware-nocturne-10984.B,master
TEST=Flash nocturne with battery stuck in disconnected state, verify
that Try.Src is not attempted.

Change-Id: I232dbb365fcdaf504b1957114bd1e29afd4cb2b2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1474570
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1603119
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>