tpm: explicitly include tpm2 data in the appropriate section
With the recent move to enabling LTO the tpm_registers module's data
variables which are supposed to be cleared on TPM reset ended up not
included in the appropriate part of the .bss section.

With this patch all variables of interest are marked with the section
name they need to be placed in.

BRANCH=cr50, cr50-mp
BUG=b:124450448
TEST=before this patch after RMA clear every TPM reset was wiping out
     the TPM NVMEM, with this patch in place only the following reset
     wipes out the NVMEM.

Change-Id: If0cbdeae2ff37f2b9cf67672c0a3b5c57efd6248
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1474892
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit 8664d3093e613b7e84fdc1c3a4b715749045064e)
Reviewed-on: https://chromium-review.googlesource.com/c/1474742