common/usb_pd_protocol: Set port in LPM at the start of init
During initialization, PD ports are not marked as in Low Power Mode.
This causes other tasks not to wait before the TCPC is initialized
leading to invalid state information. Set the ports as in Low Power Mode
at the start of initialization, so that other tasks wait until the TCPC
is initialized and ready.

BUG=chromium:934332, b:124131293
BRANCH=octopus
TEST=Ensure that the system boots to ChromeOS. Ensure that the device
detects the USB stick attached to servo v4 at recovery mode with or
without the power pass through and boots from it.

Change-Id: Id4963fc0540ab092bb5be20cada64fe34a9272de
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1492531
Commit-Ready: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1603121
Reviewed-by: Edward Hill <ecgh@chromium.org>
Commit-Queue: Edward Hill <ecgh@chromium.org>
Tested-by: Edward Hill <ecgh@chromium.org>