util: add script for comparing source files across branches
The new script allows to find all source and include files used when
building a Cr50 image and then compare all these files with a
different branch.

By default the comparison is done with cros/firmware-cr50-9308.B,
which makes it easy to compare MP and pre-pvt Cr50 branches.

BRANCH=none
BUG=none

TEST=ran the script in the Cr50 mp branch and observed the expected
     results:

 $ ./util/bcmp
 Will rebuild CR50
 Now looking for .c and .S files
 Now looking for .h files
 Comparing local tree to cros/firmware-cr50-9308.B
 ....

Change-Id: Iae3083b5775216998f7c95b43eb892054ff0c8df
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1548085
Reviewed-by: Mary Ruthven <mruthven@chromium.org>