ish: send request id back as response id
The AP needs a way to pair request and response messages as they come in
asynchronously or late. This does not help the normal case, but the error
case when the responses get behind.

BRANCH=none
BUG=b:130039715
TEST=See that kernel messages get dropped instead of error because of
mismatched size (which is a signal that responses were getting paired up
incorrectly)

Change-Id: I7ddee42aac03429e4fc6cccd47527cf4776cff04
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1560651
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1569794