Flapjack: Turn off LED when brightness == 0
Currently, when the brightness passed to led_set_brightness is zero,
the current is set to 4mA and the pwm duty is set to the brightness.
Since duty == 0 means 1/32 (instead of 0/32), this doesn't turn off
the LED.

This patch makes the current explicitly set to 0mA if the brightness
is zero.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/131205169
BRANCH=none
TEST=buildall

Change-Id: I7f581349713d30f10acb3797ab08b15aa2d50f00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1585385
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Tested-by: YH Lin <yueherngl@chromium.org>