Nami: Allow power role swap to sink in S0ix
Currently, if a hub is plugged in during S0 and a system is suspended to
S0ix, we will reject any requests from that hub to power role swap if it
becomes powered (note this is not a concern for S5, when we will
disconnect a sinking partner, so on becoming powered the hub would start
a new connection).

This change allows role swaps if we are currently sourcing a partner who
could be providing us power instead.

BUG=b:143572158
BRANCH=nami
TEST=plugged in unpowered 3-in-1 dongle to casta in S0, closed the lid
to trigger S0ix, plugged in power to the dongle and saw power role swap
was accepted

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I76eea8d74df1b827ae39d54392627e90db0d6a95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1894766