g: allow I2CS operate without hardware resets
It is not always possible to rely on PMU for resetting the I2CS
controller. Most of the AP firmware versions deploy the 'I2C unwedge'
cycle when coming out of reset, but not all of them, this is why Cr50
needs to be able to recover on its own in case there was a crash and
the I2C bus was left mid transaction with the H1 holding down the SDA
line.

A GPIO is dedicated to monitor the I2CS_SDA line during reset. If the
line is kept low, it could be a sign of a 'wedged' controller. The g
I2CS FSM will reset any time the I2C 'stop' condition is detected.

The create the 'stop' condition the I2C_SCL input is disconnected from
the bus and connected to an internal GPIO, then I2C_SCL level is set
to 'high' and register inverting the I2C_SDA value is toggled, which
looks like a transition from zero to one to the controller. thus
creating the 'stop' condition.

BRANCH=cr50, cr50-mp
BUG=b:135772657

TEST=the test was ran on a Pyro device, which uses I2C for
     communication with H1 and which AP firmware does not deploy the
     'I2C unwedge' cycle.

     Test instrumentation involved setting a Chrome OS startup file
     such that once booted, the AP starts continuously polling TPM for
     value of an NVMEM index, creating I2C traffic. The host
     workstation sends the 'apreset cold' command to the EC within a
     few seconds of Chrome OS coming up.

     First run a special Cr50 image which is not resetting I2CS using
     PMU on TPM restarts, is was not trying to unwedge the stuck I2C
     bus. On five experiments, it takes on average 32 reboots for
     until I2C bus is locked up and the DUT falls into recovery.

     Then loaded the Cr50 image with this patch and ran the test
     again, it survived for 150 cycles without a problem.

Change-Id: Iffec33f97557e3acfd1cd5fb76ba158f8c23b608
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1730143
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
(cherry picked from commit fa6f77cd0da5b2f17494201e0770d1c8d550b7f4)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1920362
(cherry picked from commit 0ccab68f150349f866528f612f55f225ba9adf98)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1959116
Tested-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>